Moved code with MessageService in MessagesToPlatform
