+++ src/main/java/com/community/tools/service/discord/MessagesToDiscord.java
@@ -86,6 +86,13 @@ public class MessagesToDiscord extends MessagesToPlatform<MessageEmbed> {
         .build();
   }
 
+  @Override
+  public MessageEmbed failedBuildMessage(String url, String task) {
+    return new EmbedBuilder()
+        .addField("", MessagesToDiscord.FAILED_BUILD + task + "](" + url + ") is down!", false)
+        .build();
+  }
+
   public static final String NEXT_TASK = "Here is your next [TASK]("
       + "https://github.com/Broscorp-net/traineeship/tree/master/module1/src/main/java/net/broscorp/";
